単語一括削除機能の不具合の修正、単語登録がない状態で一括編集削除機能をクリックした場合にflashメッセージと共に単語管理ページに戻る機能の追加

+++ app/controllers/cards_controller.rb
@@ -44,6 +44,10 @@ class CardsController < ApplicationController
         
         # 選択されたリストのid
         selected_ids= params[:selected_ids]
+        if selected_ids.nil?
+            flash[:notice] = "単語カードを選択してください"
+            redirect_to "/cards_list/#{current_user.id}"
+        end
         @selected_cards=Card.where(id: selected_ids)
         # 繰り返し処理で一つずつ格納
         # selected_ids.each do |id|
@@ -59,16 +63,22 @@ class CardsController < ApplicationController
             card = Card.find(id)
             card.update(selected_card_params("card_id_#{id}"))
         end
-        redirect_to "/cards_list/#{current_user.id}"
+        redirect_to :list
+        
     end
     
     def destroy_multiple
         # 選択されたレコードの繰り返し処理
-        # binding.pry
         ids = params[:selected_ids]
-        ids.each do |id|
-            card=Card.find(id)
-            card.destroy
+        # binding.pry
+        if ids.nil?
+            flash[:notice] = "単語カードを選択してください"
+        else
+            ids = params[:selected_ids]
+            ids.each do |id|
+                card=Card.find(id)
+                card.destroy
+            end
         end
         redirect_to "/cards_list/#{current_user.id}"
